src/requirements.txt
numpy==1.26.4  
scikit-learn==1.0.2
Werkzeug==2.0.1
Flask==2.0.3
python-dotenv==0.19.2
Flask-PyMongo==2.3.0
pandas==1.3.4
tensorflow==2.16.1  
joblib==1.1.0

